Reboot-Adjust 1.4

+++ conf/ibuntu_grub.sh
@@ -1,8 +1,27 @@
 #!/bin/bash
 # Replace OS grub with ibuntubuilder's custom grub
-cp /etc/default/grub.ibuntu /etc/default/grub
-update-grub
-systemctl disable ibuntu_grub.service
-rm /usr/share/applications/reboot.desktop
-rm /etc/systemd/system/ibuntu_grub.service
-#rm /etc/init.d/ibuntu_grub.sh
+
+count=1
+
+if [ ! -f /usr/lib/penguins-eggs/boot_count.txt ]; then
+    echo $count > /usr/lib/penguins-eggs/boot_count.txt
+fi
+
+
+count="`cat /usr/lib/penguins-eggs/boot_count.txt`"
+
+
+if [ $count -le 3 ]
+then
+   cp /etc/default/grub.ibuntu /etc/default/grub
+   update-grub
+   count=$count+1
+   echo $count > /usr/lib/penguins-eggs/boot_count.txt
+fi
+
+if [ $count -gt 3 ]
+then
+   systemctl disable ibuntu_grub.service
+   rm /etc/systemd/system/ibuntu_grub.service
+   rm /etc/init.d/ibuntu_grub.sh
+fi
